--- src/main/java/at/hannibal2/skyhanni/utils/tracker/DisplayMode.kt
@@ -1,7 +1,0 @@
-package at.hannibal2.skyhanni.utils.tracker
-
-enum class DisplayMode(val displayName: String) {
-    TOTAL("Total"),
-    SESSION("This Session"),
-    ;
-}
